Refactor: Frame
